normalize signature treatment
