+++ apps/broker-service/src/signedTimestamp.ts
@@ -1,17 +1,17 @@
+import { DateTime } from 'luxon';
 import { SignedMessage } from '@thecointech/types';
 import { GetSigner } from '@thecointech/utilities/SignedMessages';
+import { assertFresh } from './assertSigner';
 
 //
 // Check the packet to ensure it's valid
 // Returns address of signer if successful
-export function getSigner(sm: SignedMessage) {
-  // First, valid message?
-	const ts = parseInt(sm.message);
-	// Message should be timestamp, within the last 5 minutes
-	const age = Date.now() - ts;
-	if (age > (5 * 60 * 1000))
-		throw new Error("Timestamp too old");
+export async function getSigner(sm: SignedMessage) {
+  const signer = await GetSigner(sm);
 
-  // Ok - it's a valid message.  Get the signer
-  return GetSigner(sm);
+  // Message should be a timestamp, signed within the last 5 minutes
+  const signedAt = DateTime.fromMillis(parseInt(sm.message));
+  assertFresh(signedAt, 'verify signed timestamp', signer);
+
+  return signer;
 }
